+++ src/components/common/Sign/SignForm.tsx
@@ -1,7 +1,6 @@
-import { useState } from 'react'
-import { Link, useNavigate } from 'react-router-dom'
+import { Link } from 'react-router-dom'
 
-import { SignFormProps, SigninResponse } from './types'
+import { SignFormProps } from './types'
 import {
   StyledSignFormContainer,
   StyledSignFormTitle,
@@ -11,57 +10,12 @@ import {
   StyledSignFormLinkWrap,
   StyledSignFormButton,
 } from './SignForm.styled'
-import { postSignup, postSignin } from '../../../apis/auth'
-import { useAuthContext } from '../../../AuthProvider'
+import useForm from '../../../hooks/useForm'
 
 function SignForm({ isSignUp }: SignFormProps) {
-  const navigate = useNavigate()
-  const { updateAuth } = useAuthContext()
-  const [email, setEmail] = useState('')
-  const [password, setPassword] = useState('')
-  const [isValidEmail, setIsValidEmail] = useState(false)
-  const [isValidPassword, setIsValidPassword] = useState(false)
-
-  const onChangeEmail = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const value = e.target.value
-    setEmail(value)
-    setIsValidEmail(/@/.test(value))
-  }
-
-  const onChangePassword = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const value = e.target.value
-    setPassword(value)
-    setIsValidPassword(/.{8,}$/.test(value))
-  }
-
-  const sendData = () => {
-    const data = {
-      email: email,
-      password: password,
-    }
-    if (isSignUp) {
-      postSignup(data.email, data.password)
-        .then(() => {
-          alert('회원가입 완료!')
-          navigate('/signin')
-        })
-        .catch((err) => {
-          const message = err.response.data.message
-          alert(message)
-          console.error(err)
-        })
-    } else {
-      postSignin(data.email, data.password)
-        .then((res: SigninResponse) => {
-          updateAuth({ action: 'sign-in', tokenValue: res.data.access_token })
-        })
-        .catch((err) => {
-          const message = err.response.data.message
-          alert(message)
-          console.error(err)
-        })
-    }
-  }
+  const { onChangeEmail, onChangePassword, sendData, isValidEmail, isValidPassword } = useForm({
+    isSignUp,
+  })
 
   return (
     <StyledSignFormContainer>
